backend/analytics/requirements.txt
scikit-learn
pandas
numpy
mysql-connector-python
statsmodels
json
sys
argparse
